feat: 연관관계 구현 및 엔티티 어노테이션 수정
연관관계 수정 및 엔티티 어노테이션 수정

+++ back/src/main/java/com/mainproject/back/follow/entity/Follow.java
@@ -25,7 +25,7 @@ public class Follow extends Auditable {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int followId;
+  private Long followId;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "follower_id")
